Bug fixes (customer, order)

+++ src/main/java/com/northwind/northwindrestapi/service/CustomerService.java
@@ -1,10 +1,13 @@
 package com.northwind.northwindrestapi.service;
 
 import com.northwind.northwindrestapi.dao.CustomerRepository;
+import com.northwind.northwindrestapi.dto.CustomerCreateDTO;
 import com.northwind.northwindrestapi.dto.CustomerPatchDTO;
 import com.northwind.northwindrestapi.entity.Customer;
+import com.northwind.northwindrestapi.mapper.CustomerMapper;
 import com.northwind.northwindrestapi.service.interfaces.ICustomerService;
 import lombok.AllArgsConstructor;
+import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -16,6 +19,7 @@ import java.util.List;
 public class CustomerService implements ICustomerService {
 
   private CustomerRepository customerRepository;
+  private CustomerMapper customerMapper;
 
   @Override
   public List<Customer> getAllCustomers() {
@@ -24,24 +28,33 @@ public class CustomerService implements ICustomerService {
 
   @Override
   public Customer getCustomer(String id) {
-    return customerRepository.findById(id).orElse(null);
+    return customerRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
   }
 
   @Override
-  public Customer create(Customer customer) {
-    return customerRepository.saveAndFlush(customer);
+  public Customer create(CustomerCreateDTO customer) {
+    String id = RandomStringUtils.random(5, true, false);
+    Customer entity = customerMapper.createDtoToEntity(customer);
+    entity.setId(id);
+    return customerRepository.saveAndFlush(entity);
   }
 
   @Override
-  public Customer update(String id, Customer customer) {
+  public Customer update(String id, CustomerCreateDTO customer) {
     if (customerRepository.existsById(id)) {
-      return customerRepository.saveAndFlush(customer);
+      Customer entity = customerMapper.createDtoToEntity(customer);
+      entity.setId(id);
+      return customerRepository.saveAndFlush(entity);
     }
-    else return null;
+    else
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND);
   }
 
   @Override
   public void delete(String id) {
+    if (!customerRepository.existsById(id)) {
+      throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+    }
     customerRepository.deleteById(id);
   }
